requirements.txt
edge_tts==7.2.7
fastapi==0.135.1
groq==1.1.0
psycopg2_binary==2.9.9
pydantic==2.12.5
python-dotenv==1.2.2
scikit_learn==1.7.2
SQLAlchemy==2.0.23
uvicorn
asyncpg
pydantic[email]
python-multipart